Get a list of unique sectors

These include which sector the underlying data belongs to. This will be helpful while querying for the right API.

get_list_of_sectors()

Value

character vector of sectors

Examples

get_list_of_sectors()
#>   [1] "Environment and Forest"               
#>   [2] "Vehicular Air Pollution"              
#>   [3] "Residential Air Pollution"            
#>   [4] "Industrial Air Pollution"             
#>   [5] "Power and Energy"                     
#>   [6] "Petroleum and Natural Gas"            
#>   [7] "Health and Family welfare"            
#>   [8] "Health"                               
#>   [9] "All"                                  
#>  [10] "Water and Sanitation"                 
#>  [11] "Industries"                           
#>  [12] "Medium"                               
#>  [13] "Micro"                                
#>  [14] "Small Scale"                          
#>  [15] "Agriculture"                          
#>  [16] "Agricultural Marketing"               
#>  [17] "Family Welfare"                       
#>  [18] "Finance"                              
#>  [19] "Economy"                              
#>  [20] "Statistics"                           
#>  [21] "Census and Surveys"                   
#>  [22] "Census"                               
#>  [23] "Population Enumerator"                
#>  [24] "Development"                          
#>  [25] "Education"                            
#>  [26] "Elementary"                           
#>  [27] "Secondary"                            
#>  [28] "Non Renewable"                        
#>  [29] "Home Affairs and Enforcement"         
#>  [30] "Police"                               
#>  [31] "Macro Economy"                        
#>  [32] ""                                     
#>  [33] "Post"                                 
#>  [34] "Manufacturing"                        
#>  [35] "Transport"                            
#>  [36] "Aviation"                             
#>  [37] "Textiles"                             
#>  [38] "Insurance"                            
#>  [39] "Governance and Administration"        
#>  [40] "Fisheries"                            
#>  [41] "Prices"                               
#>  [42] "Chemicals and Petrochemicals"         
#>  [43] "Surface Water"                        
#>  [44] "Water Resources"                      
#>  [45] "Science and Technology"               
#>  [46] "Information and Communications"       
#>  [47] "Information and Technology"           
#>  [48] "Road Transport"                       
#>  [49] "Horticulture"                         
#>  [50] "Mining"                               
#>  [51] "Employment"                           
#>  [52] "Water ways"                           
#>  [53] "Commerce"                             
#>  [54] "Companies"                            
#>  [55] NA                                     
#>  [56] "Animal Husbandry"                     
#>  [57] "Urban"                                
#>  [58] "Earth Sciences"                       
#>  [59] "Agricultural Produces"                
#>  [60] "Travel and Tourism"                   
#>  [61] "Telecom"                              
#>  [62] "Atmospheric Science"                  
#>  [63] "Export"                               
#>  [64] "Import"                               
#>  [65] "Labour and Employment"                
#>  [66] "Food"                                 
#>  [67] "Heavy"                                
#>  [68] "Internal Security"                    
#>  [69] "Others"                               
#>  [70] "Noise Pollution"                      
#>  [71] "Sanitation"                           
#>  [72] "House listing and Housing Census"     
#>  [73] "Revenue"                              
#>  [74] "Census and Surveys – ORGI"            
#>  [75] "Local Government"                     
#>  [76] "Higher Education"                     
#>  [77] "Infrastructure"                       
#>  [78] "Power"                                
#>  [79] "Housing"                              
#>  [80] "Urban Housing"                        
#>  [81] "Renewable"                            
#>  [82] "Drinking Water"                       
#>  [83] "District Adminstration"               
#>  [84] "National Population Register"         
#>  [85] "Municipal Waste"                      
#>  [86] "Roads"                                
#>  [87] "Water Quality"                        
#>  [88] "Research & Development"               
#>  [89] "Banking"                              
#>  [90] "Agricultural Research & Extension"    
#>  [91] "Ground Water"                         
#>  [92] "Art and Culture"                      
#>  [93] "Railways"                             
#>  [94] "UrbanLocal Government"                
#>  [95] "Adult Education"                      
#>  [96] "Dairying"                             
#>  [97] "Land Resources"                       
#>  [98] "Fertilizers"                          
#>  [99] "Social Development"                   
#> [100] "Rural"                                
#> [101] "Fishery"                              
#> [102] "Panchayati Raj"                       
#> [103] "Broadcasting"                         
#> [104] "Information and Broadcasting"         
#> [105] "Defence"                              
#> [106] "Rajya Sabha"                          
#> [107] "Parliament Of india"                  
#> [108] "Pension Reforms"                      
#> [109] "Lok Sabha"                            
#> [110] "Public Distribution"                  
#> [111] "Crops"                                
#> [112] "Tribal"                               
#> [113] "Union/State Government Administration"
#> [114] "Natural Resources"                    
#> [115] "Passport"                             
#> [116] "Tourism"                              
#> [117] "Civil Registration System"            
#> [118] "Lodging"                              
#> [119] "Biotechnology"                        
#> [120] "Seeds"                                
#> [121] "Judiciary"                            
#> [122] "Agricultural"                         
#> [123] "Film"                                 
#> [124] "Sericulture"                          
#> [125] "ElementaryEducation"                  
#> [126] "CensusCensus and Surveys - ORGI"      
#> [127] "Visa"                                 
#> [128] "Foreign Affairs"                      
#> [129] "Heritage"                             
#> [130] "Consumer Affairs"