Search API by organization name

Search API by organization name

search_api_by_organization(organization_name_contains = "")

Arguments

organization_name_contains

string to match in the organization type of the API (case insensitive)

Value

data.frame of API details filtered by organization names (Ministries) that match the string provided

Examples

  get_list_of_organizations()
#>   [1] "Ministry of Environment, Forest and Climate Change"                                                    
#>   [2] "Central Pollution Control Board"                                                                       
#>   [3] "Ministry of Power"                                                                                     
#>   [4] "Ministry of Petroleum and Natural Gas"                                                                 
#>   [5] "Ministry of Health and Family Welfare"                                                                 
#>   [6] "Department of Health and Family Welfare"                                                               
#>   [7] "Central Government Health Scheme (CGHS)"                                                               
#>   [8] "Gujarat"                                                                                               
#>   [9] "Surat Municipal Corporation"                                                                           
#>  [10] "Ministry of Jal Shakti"                                                                                
#>  [11] "Department of Drinking Water and Sanitation"                                                           
#>  [12] "Ministry of Micro, Small and Medium Enterprises"                                                       
#>  [13] "Ministry of Agriculture and Farmers Welfare"                                                           
#>  [14] "Department of Agriculture and Farmers Welfare"                                                         
#>  [15] "Ministry of Statistics and Programme Implementation"                                                   
#>  [16] "Directorate of Marketing and Inspection (DMI)"                                                         
#>  [17] "Ministry of Home Affairs"                                                                              
#>  [18] "Department of Home"                                                                                    
#>  [19] "Registrar General and Census Commissioner, India"                                                      
#>  [20] "Karnataka"                                                                                             
#>  [21] "Urban Development Department, Karnataka"                                                               
#>  [22] "Karnataka Municipal Data Society"                                                                      
#>  [23] "Department of School Education and Literacy"                                                           
#>  [24] "Ministry of Education"                                                                                 
#>  [25] "Ministry of Coal"                                                                                      
#>  [26] "Rajya Sabha"                                                                                           
#>  [27] "Department of States"                                                                                  
#>  [28] "National Crime Records Bureau (NCRB)"                                                                  
#>  [29] "Ministry of Communications"                                                                            
#>  [30] "Department of Posts"                                                                                   
#>  [31] "Ministry of Commerce and Industry"                                                                     
#>  [32] "Department of Industrial Policy and Promotion"                                                         
#>  [33] "Ministry of Civil Aviation"                                                                            
#>  [34] "Department of School Education & Literacy"                                                             
#>  [35] "NITI Aayog"                                                                                            
#>  [36] "Home Department, Karnataka"                                                                            
#>  [37] "Directorate of Home Guards and Civil Defense"                                                          
#>  [38] "Department of Skill Development Entrepreneurship and Livelihood"                                       
#>  [39] "Government Tool Room And Training Center(GTTC)"                                                        
#>  [40] "Health and Family Welfare Department, Karnataka"                                                       
#>  [41] "Department of Ayush"                                                                                   
#>  [42] "Planning and Statistics Department, Karnataka"                                                         
#>  [43] "Directorate of Economics & Statistics"                                                                 
#>  [44] "Ministry of Finance"                                                                                   
#>  [45] "Department of Financial Services"                                                                      
#>  [46] "Energy Department, Karnataka"                                                                          
#>  [47] "Department of Personnel and Administrative Reforms, Karnataka"                                         
#>  [48] "Karnataka Sakala Services"                                                                             
#>  [49] "Animal Husbandry and Fisheries Department, Karnataka"                                                  
#>  [50] "Central Electricity Authority"                                                                         
#>  [51] "Ministry of Chemicals and Fertilizers"                                                                 
#>  [52] "Department of Chemicals and Petrochemicals"                                                            
#>  [53] "Department of Water Resources, River Development & Ganga Rejuvenation"                                 
#>  [54] "National Water Informatics Centre"                                                                     
#>  [55] "Tamil Nadu"                                                                                            
#>  [56] "Health and Family Welfare Department"                                                                  
#>  [57] "Public Health and Preventive Medicine"                                                                 
#>  [58] "Ministry of Science and Technology"                                                                    
#>  [59] "Ministry of Electronics and Information Technology (MeitY)"                                            
#>  [60] "Department of Economic Affairs"                                                                        
#>  [61] "Department of Revenue"                                                                                 
#>  [62] "Financial Intelligence Unit - India"                                                                   
#>  [63] "Ministry of Mines"                                                                                     
#>  [64] "Indian Bureau of Mines"                                                                                
#>  [65] "Meghalaya"                                                                                             
#>  [66] "Ministry of Ports, Shipping and Waterways"                                                             
#>  [67] "Ministry of Corporate Affairs"                                                                         
#>  [68] "Department for Promotion of industry and Internal Trade"                                               
#>  [69] NA                                                                                                      
#>  [70] "Ministry of Law and Justice"                                                                           
#>  [71] "Department of Legal Affairs"                                                                           
#>  [72] "Department of Public Enterprises"                                                                      
#>  [73] "Ministry of Fisheries, Animal Husbandry and Dairying"                                                  
#>  [74] "Department of Animal Husbandry and Dairying"                                                           
#>  [75] "Tiruppur"                                                                                              
#>  [76] "Ministry of Steel"                                                                                     
#>  [77] "Planning, Development and Special Initiatives Department"                                              
#>  [78] "Department of Economics and Statistics"                                                                
#>  [79] "Ministry of Earth Sciences"                                                                            
#>  [80] "India Meteorological Department (IMD)"                                                                 
#>  [81] "India Meteorological Department (IMD), Pune"                                                           
#>  [82] "Ministry of Tourism"                                                                                   
#>  [83] "Department of Telecommunications (DOT)"                                                                
#>  [84] "Telecom Regulatory Authority of India (TRAI)"                                                          
#>  [85] "National Health Mission, Tamil Nadu"                                                                   
#>  [86] "Department of Commerce"                                                                                
#>  [87] "Agriculture Department"                                                                                
#>  [88] "Horticulture & Plantation Crops"                                                                       
#>  [89] "Tamil Nadu Medical Services Corporation (TNMSC)"                                                       
#>  [90] "Surat"                                                                                                 
#>  [91] "Information Technology Department"                                                                     
#>  [92] "Electronics Corporation of Tamilnadu Ltd (ELCOT)"                                                      
#>  [93] "Ministry of Road Transport and Highways"                                                               
#>  [94] "Ministry of Consumer Affairs, Food and Public Distribution"                                            
#>  [95] "Department of Food and Public Distribution"                                                            
#>  [96] "Food Corporation of India (FCI)"                                                                       
#>  [97] "Labour and Employment Department"                                                                      
#>  [98] "Tamil Nadu Skill Development Corporation(TNSDC)"                                                       
#>  [99] "Industries Department"                                                                                 
#> [100] "Tamil Nadu Industrial Investment Corporation (TIIC)"                                                   
#> [101] "Ministry of Agriculture and Farmers Welfare;Department of Agriculture, Cooperation and Farmers Welfare"
#> [102] "Home, Prohibition and Excise Department"                                                               
#> [103] "Tamil Nadu Fire and Rescue Services"                                                                   
#> [104] "Madhya Pradesh"                                                                                        
#> [105] "Indore"                                                                                                
#> [106] "Madurai"                                                                                               
#> [107] "Tirunelveli"                                                                                           
#> [108] "Rajasthan"                                                                                             
#> [109] "Kota"                                                                                                  
#> [110] "Davangere"                                                                                             
#> [111] "Haryana"                                                                                               
#> [112] "Faridabad"                                                                                             
#> [113] "Maharashtra"                                                                                           
#> [114] "Nagpur"                                                                                                
#> [115] "Belagavi"                                                                                              
#> [116] "Andhra Pradesh"                                                                                        
#> [117] "Tirupati"                                                                                              
#> [118] "Vadodara"                                                                                              
#> [119] "Punjab"                                                                                                
#> [120] "Amritsar"                                                                                              
#> [121] "Chhattisgarh"                                                                                          
#> [122] "Naya Raipur"                                                                                           
#> [123] "Thane"                                                                                                 
#> [124] "Uttar Pradesh"                                                                                         
#> [125] "Kanpur"                                                                                                
#> [126] "Pimpri Chinchwad"                                                                                      
#> [127] "Nashik"                                                                                                
#> [128] "Nagaland"                                                                                              
#> [129] "Kohima"                                                                                                
#> [130] "Raipur"                                                                                                
#> [131] "Bilaspur"                                                                                              
#> [132] "Varanasi"                                                                                              
#> [133] "Aurangabad"                                                                                            
#> [134] "Visakhapatnam"                                                                                         
#> [135] "Bareilly"                                                                                              
#> [136] "Coimbatore"                                                                                            
#> [137] "Chennai"                                                                                               
#> [138] "Odisha"                                                                                                
#> [139] "Rourkela"                                                                                              
#> [140] "Tiruchirapalli"                                                                                        
#> [141] "Thanjavur"                                                                                             
#> [142] "Thoothukudi"                                                                                           
#> [143] "Mangaluru"                                                                                             
#> [144] "Salem"                                                                                                 
#> [145] "Jhansi"                                                                                                
#> [146] "Udaipur"                                                                                               
#> [147] "Himachal Pradesh"                                                                                      
#> [148] "Dharamshala"                                                                                           
#> [149] "Mizoram"                                                                                               
#> [150] "Aizawl"                                                                                                
#> [151] "Kerala"                                                                                                
#> [152] "Tiruvanantapuram"                                                                                      
#> [153] "National Institute of Health and Family Welfare (NIHFW), New Delhi"                                    
#> [154] "Ahmedabad"                                                                                             
#> [155] "Pune"                                                                                                  
#> [156] "Gwalior"                                                                                               
#> [157] "Chandigarh"                                                                                            
#> [158] "Agra"                                                                                                  
#> [159] "Rajkot"                                                                                                
#> [160] "Hubli Dharwad"                                                                                         
#> [161] "Tripura"                                                                                               
#> [162] "Agartala"                                                                                              
#> [163] "Telangana"                                                                                             
#> [164] "Karimnagar"                                                                                            
#> [165] "Shimla"                                                                                                
#> [166] "Bihar"                                                                                                 
#> [167] "Patna"                                                                                                 
#> [168] "Tumakuru"                                                                                              
#> [169] "Jaipur"                                                                                                
#> [170] "Jammu and Kashmir"                                                                                     
#> [171] "Jammu"                                                                                                 
#> [172] "Shillong"                                                                                              
#> [173] "Andaman & Nicobar Islands"                                                                             
#> [174] "Port Blair"                                                                                            
#> [175] "Bhopal"                                                                                                
#> [176] "Bengaluru"                                                                                             
#> [177] "Kakinada"                                                                                              
#> [178] "Amravati"                                                                                              
#> [179] "Ajmer"                                                                                                 
#> [180] "Bihar Sharif"                                                                                          
#> [181] "Delhi"                                                                                                 
#> [182] "NDMC"                                                                                                  
#> [183] "Greater Warangal"                                                                                      
#> [184] "West Bengal"                                                                                           
#> [185] "New Town Kolkata"                                                                                      
#> [186] "Department of Higher Education"                                                                        
#> [187] "Department of Science and Technology (DST)"                                                            
#> [188] "National Science and Technology Management Information System (NSTMIS)"                                
#> [189] "Department of Agricultural Research and Education (DARE)"                                              
#> [190] "Indian Council of Agricultural Research (ICAR)"                                                        
#> [191] "Department of Atomic Energy"                                                                           
#> [192] "Public Works Department , Tamil Nadu"                                                                  
#> [193] "Planning, Development and Special Initiatives Department, Tamil Nadu"                                  
#> [194] "Energy Department, Tamil Nadu"                                                                         
#> [195] "Industries Department, Tamil Nadu"                                                                     
#> [196] "Handlooms, Handicrafts and Textiles Department, Tamil Nadu"                                            
#> [197] "Handlooms and Textiles Directorate"                                                                    
#> [198] "Tourism, Culture and Religious Endowments Department"                                                  
#> [199] "Transport Department, Tamil Nadu"                                                                      
#> [200] "Health and Family Welfare Department, Tamil Nadu"                                                      
#> [201] "Directorate of Medical Education"                                                                      
#> [202] "Municipal Administration and Water Supply Department, Tamil Nadu"                                      
#> [203] "Micro Small and Medium Enterprises Department, Tamil Nadu"                                             
#> [204] "Tamil Nadu Small Industries Development Corporation Ltd (TANSIDCO)"                                    
#> [205] "Information Technology Department, Tamil Nadu"                                                         
#> [206] "Tamil Nadu e-Governance Agency (TNEGA)"                                                                
#> [207] "Ministry of Railways"                                                                                  
#> [208] "Commissionerate of Municipal Administration"                                                           
#> [209] "National Council of Educational Research and Training (NCERT)"                                         
#> [210] "Ministry of Rural Development"                                                                         
#> [211] "Department of Land Resources (DLR)"                                                                    
#> [212] "Department of Fertilizers"                                                                             
#> [213] "Office of the Economic Adviser"                                                                        
#> [214] "Ministry of Panchayati Raj"                                                                            
#> [215] "Ministry of Information and Broadcasting"                                                              
#> [216] "Economic and Statistical Analysis Department, Haryana"                                                 
#> [217] "Planning, Economics and Statistics Department, Madhya Pradesh"                                         
#> [218] "Ministry of Defence"                                                                                   
#> [219] "Department of Defence Research and Development"                                                        
#> [220] "Defence Research and Development Organisation (DRDO)"                                                  
#> [221] "Food, Civil Supplies and Consumer Protection Department, Madhya Pradesh"                               
#> [222] "Transport Department, Madhya Pradesh"                                                                  
#> [223] "Directorate General of Foreign Trade (DGFT)"                                                           
#> [224] "Animal Husbandary, Darirying and Fisheries Department, Tamil Nadu"                                     
#> [225] "School Education Department, Tamil Nadu"                                                               
#> [226] "Sarva Sihiksha Abhiyan"                                                                                
#> [227] "Ministry of Tribal Affairs"                                                                            
#> [228] "Delhi (NCT)"                                                                                           
#> [229] "Health and Family Welfare Department, Delhi"                                                           
#> [230] "Lal Bahadur Shastri Hospital"                                                                          
#> [231] "Information Technology Department, Madhya Pradesh"                                                     
#> [232] "Ministry of External Affairs"                                                                          
#> [233] "Ministry of Development of North Eastern Region"                                                       
#> [234] "Department of Biotechnology, Government of India"                                                      
#> [235] "Commerce, Industry and Employment Department, Madhya Pradesh"                                          
#> [236] "Co-operation Department, Madhya Pradesh"                                                               
#> [237] "Lok Sabha Secretariat"                                                                                 
#> [238] "Environment and Forests Department, Tamil Nadu"                                                        
#> [239] "Ministry of Heavy Industries and Public Enterprises"                                                   
#> [240] "Department of Heavy Industries"                                                                        
#> [241] "Ministry of AYUSH"                                                                                     
#> [242] "Excise, Registration, Taxation and Stamps (ERTS) Department, Meghalaya"                                
#> [243] "Department of AIDS Control"                                                                            
#> [244] "Adi Dravidar and Tribal Welfare Department, Tamil Nadu"                                                
#> [245] ""                                                                                                      
#> [246] "Ministry of New and Renewable Energy"                                                                  
#> [247] "Farmer Welfare and Agriculture Development Department, Madhya Pradesh"                                 
#> [248] "Agriculture Department, Meghalaya"                                                                     
#> [249] "Department of Consumer Affairs"                                                                        
filtered_api<-search_api_by_organization(organization_name_contains="Agriculture")